move concat macros to orkstd. move spscque to concurrent_queue. fix logic with overflow calc on concurrent_queue.
